requirements.txt
fastapi
uvicorn
python-dotenv
python-multipart
langchain-openai
langchain
langchain-core
pypdf
